feat: enhance FFI with output metadata handling for shape operations
- Updated various shape operations (flatten, ravel, insert_axis, invert_axis, merge_axes, pad, permute_axes, squeeze, swap_axes, transpose, concatenate, stack) to include output metadata handling.
- Added checks for null pointers in output metadata parameters to ensure robust error handling.
- Enhanced PHP bindings to accommodate new output metadata parameters for relevant functions.

+++ tests/Unit/ReductionsTest.php
@@ -115,7 +115,7 @@ final class ReductionsTest extends TestCase
 
     public function testSumAxis1(): void
     {
-        $a = NDArray::array([[1, 2, 3], [4, 5, 6]], DType::Float64);
+        $a = NDArray::array([[1, 2, 3], [4, 5, 6]], DType::Int16);
         $result = $a->sum(axis: 1);
         $this->assertEqualsWithDelta([6, 15], $result->toArray(), 0.0001);
         $this->assertSame([2], $result->shape());
